fix: fix regex error in title processing subroutine
This commit is reverted from other commit which included somewhat clumsy OOP flavor into the code. It enclosed every implemented method in the preprocess module into a class. That was overdone already since everything worked fine without class wrapping, but it got worse after adding more abstraction with interface implementation, even if there was no room for any polymorphism.

It is good to remember that main characteristic of Python is its high productivity coming from somewhat loose grammatical or structural restrictions(remember, it is dynamically typed language welcoming duck typing!). So if something can be implemented with lesser code, it is best to do so for the sake of using such low-performance language.